Separating Things to a header
