+++ src/makefile.main.Rt
@@ -37,7 +37,7 @@ SOURCE_PLAN+=compare.cpp
 SOURCE_PLAN+=simplepart.cpp
 SOURCE_PLAN+=GetThreads.h GetThreads.cpp
 SOURCE_PLAN+=range_int.hpp
-SOURCE_PLAN+=Lists.h Lists.cpp
+SOURCE_PLAN+=Lists.h Lists.cpp Things.h
 <?R
 	h = dir("src/Handlers","[.](h|cpp)(|.Rt)$")
 	h = sub(".Rt","",h)
